Added padding to CRC16::calculation return value if it is less that than 4 character length.
During stress testing, it was observed that some QR codes failed to read. Upon investigation, it was found that the CRC16::calculate() function returned only three characters instead of the expected four, particularly when the first character was supposed to be 0. To address this issue, the padLeft function was utilized, successfully resolving the problem with broken QR codes.